+++ Sources/DynamicButtonStyles/DynamicButtonStyleHamburger.swift
@@ -29,13 +29,13 @@ import UIKit
 /// Hamburger button style: ≡
 final public class DynamicButtonStyleHamburger: DynamicButtonStyle {
   convenience required public init(center: CGPoint, size: CGFloat, offset: CGPoint, lineWidth: CGFloat) {
-    let halfSize = size / 2
+    let midSize = size / 2
 
-    let p1 = PathHelper.line(atCenter: center, radius: halfSize, angle: 0)
-    let p2 = PathHelper.line(atCenter: center, radius: halfSize, angle: 0, offset: CGPoint(x: 0, y: size / -3.2))
-    let p3 = PathHelper.line(atCenter: center, radius: halfSize, angle: 0, offset: CGPoint(x: 0, y: size / 3.2))
+    let p1 = PathHelper.line(atCenter: center, radius: midSize, angle: 0, offset: CGPoint(x: 0, y: size / -3.2))
+    let p2 = PathHelper.line(atCenter: center, radius: midSize, angle: 0)
+    let p3 = PathHelper.line(atCenter: center, radius: midSize, angle: 0, offset: CGPoint(x: 0, y: size / 3.2))
 
-    self.init(pathVector: (p1, p2, p3, p3))
+    self.init(pathVector: (p1, p2, p3, p2))
   }
 
   // MARK: - Conforming the CustomStringConvertible Protocol